requirements.txt
opencv-python
tensorflow-gpu
numpy
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
